--- schema/specs-info.json
@@ -1,21 +1,0 @@
-{
-  "$schema": "http://json-schema.org/schema#",
-  "$id": "https://github.com/w3c/browser-specs/tree/master/schema/specs-info.json",
-
-  "type": "object",
-  "propertyNames": {
-    "pattern": "^[\\w\\-]+((?<=\\-\\d+)\\.\\d+)?$"
-  },
-  "additionalProperties": {
-    "type": "object",
-    "properties": {
-      "edUrl": { "$ref": "definitions.json#/proptype/url" },
-      "trUrl": { "$ref": "definitions.json#/proptype/url" },
-      "title": { "$ref": "definitions.json#/proptype/title" },
-      "source": { "$ref": "definitions.json#/proptype/source" }
-    },
-    "required": ["title", "source"],
-    "additionalProperties": false
-  },
-  "minItems": 1
-}
