--- specs-info.json
@@ -1,7 +1,0 @@
-{
-  "compat": {
-    "edUrl": "https://compat.spec.whatwg.org/",
-    "title": "Compatibility Standard",
-    "source": "specref"
-  }
-}
